Add decimal calculation

+++ Calculator_WPF/Calculator/Calculator.cs
@@ -4,22 +4,22 @@ namespace CalculatorLib
 {
     public class Calculator
     {
-        public static int Add(int a, int b)
+        public static double Add(double a, double b)
         {
             return a + b;
         }
 
-        public static int Subtract(int a, int b)
+        public static double Subtract(double a, double b)
         {
             return a - b;
         }
 
-        public static int Multiply(int a, int b)
+        public static double Multiply(double a, double b)
         {
             return a * b;
         }
 
-        public static int Divide(int a, int b)
+        public static double Divide(double a, double b)
         {
             try
             {
